fix the revert of sinc_resampler auto spanification
spanification: automatically spanify media/base/sinc_resampler.cc etc.

This is the result of running the automatic spanification on linux and
updating code to use and pass spans where size is known.

The original patch was fully automated using script:
//tools/clang/spanify/rewrite-multiple-platforms.sh -platforms=linux
Then refined with gemini-cli

gemini-run/batch-run-1761116551/group_125

BUG=439964610

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7264486
Commit-Queue: Daniel Angulo <angdaniel@google.com>
Cr-Commit-Position: refs/heads/main@{#1565299}

+++ chromecast/media/cma/backend/mixer/mixer_input.cc
@@ -77,10 +77,10 @@ MixerInput::MixerInput(Source* source, FilterGroup* filter_group)
       // Round up to nearest multiple of SincResampler::kMaxKernelSize. The read
       // size must be > kMaxKernelSize, so we round up to at least 2 *
       // kMaxKernelSize.
-      source_read_size_ =
-          RoundUpMultiple(std::max(source_->desired_read_size(),
-                                   ::media::SincResampler::kMaxKernelSize + 1),
-                          ::media::SincResampler::kMaxKernelSize);
+      source_read_size_ = RoundUpMultiple(
+          std::max(static_cast<size_t>(source_->desired_read_size()),
+                   ::media::SincResampler::kMaxKernelSize + 1),
+          ::media::SincResampler::kMaxKernelSize);
     }
     resample_ratio_ = static_cast<double>(input_samples_per_second_) /
                       output_samples_per_second_;
